pyproject.toml
[project]
name = "zignolz-analytics-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiofiles>=24.1.0",
    "aiohttp>=3.10.0",
    "fastapi[standard]>=0.116.1",
    "jupyter>=1.1.1",
    "lxml>=6.0.1",
    "pandas>=2.3.2",
    "pandas-stubs>=2.3.2.250827",
    "polars>=1.33.1",
    "pytest>=8.4.2",
    "pytest-asyncio>=0.24.0",
    "pyyaml>=6.0",
    "requests>=2.32.0",
    "ruff>=0.13.0",
    "types-pyyaml>=6.0.12.20250822",
    "types-requests>=2.32.4.20250913",
    "yfinance>=0.2.65",
]

[dependency-groups]
dev = ["pre-commit>=4.3.0", "ruff>=0.13.0"]
